old/requirements.txt
flask==3.0.3
Flask-Babel==4.0.0
openai>=1.68.0
markdown>=3.7
SQLAlchemy>=2.0.29
python-dotenv>=1.1.0
pydantic_settings>=2.9.1
colorama>=0.4.6